Updated rating to double instead of Int

+++ app/src/main/java/com/example/finalproject/database/Day.kt
@@ -11,7 +11,7 @@ data class Day(
     val date: String,
 
     @ColumnInfo(name = "rating")
-    val rating: Int = 0,
+    val rating: Double = 0.0,
 
     @ColumnInfo(name = "journal")
     val journal: String = "",
